+++ packages/network/example/main.dart
@@ -2,8 +2,10 @@ import 'package:network/network.dart' as network;
 
 main() async {
   final getResponse = await network.get<network.JsonApiResponse>(
-      'https://jsonplaceholder.typicode.com/todos/1');
-  print(getResponse.toMap['title']);
+      'https://jsonplaceholder.typicode.com/comments', queryParameters: {
+        'postId': 1
+      });
+  print(getResponse.toList[1]['body']);
 
   // Post request to api
   final postResponse = await network.post<network.JsonApiResponse>(