+++ packages/network/lib/src/methods/get.dart
@@ -3,10 +3,15 @@ import 'package:http/http.dart' as http show get, Response;
 import 'package:network/src/exception.dart';
 import 'package:network/src/response.dart';
 import 'package:network/src/utils.dart';
+import 'package:network/src/utils/serialize_query_params.dart';
 
-Future<T> get<T extends BinaryResponse>(String url,
-    {Map<String, String> headers}) async {
-  final http.Response httpResponse = await http.get(url, headers: headers);
+Future<T> get<T extends BinaryResponse>(
+  String url, {
+  Map<String, String> headers,
+  Map<String, dynamic> queryParameters = const {},
+}) async {
+  final http.Response httpResponse = await http
+      .get(url + serializeQueryParameters(queryParameters), headers: headers);
 
   final int statusCode = httpResponse.statusCode;