+++ packages/network/CHANGELOG.md
@@ -3,6 +3,11 @@
 [comment]: <> (### Breaking Changes or ### New Features)
 [comment]: <> (* Change description)
 
+## 0.1.0
+
+* Added parameter `queryParameters` for all methods
+* Removed parameter `jsonBody` from *post* and *put* methods
+
 ## 0.0.2
 
 * Removed debug print